Consolidate capture scripts into capture.py worker + capture_driver.py
